safari-534.53-branch/LayoutTests/fullscreen/full-screen-keyboard-disabled-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitIsFullScreen == 'true') OK
EVENT(keypress)
EXPECTED (event.keyCode == '32') OK
END OF TEST
This tests that full screen mode does not allow alphanumeric key events. Press space to continue, then press an alphanumeric key, followed by space.
